server organized and working - able to upload new posts to the server. note: make sure to change the cors origin to your localhost because it was 5174 before and now it's 5173, so I had to change it
